Move Section class to separate file
