+++ server/src/validationRules/colorsThresholds.ts
@@ -1,7 +1,8 @@
 import { Diagnostic } from "vscode-languageserver";
-import { RelatedSettingsRule, Section } from "../configTree";
+import { RelatedSettingsRule } from "../configTree";
 import { incorrectColors } from "../messageUtil";
 import { sectionMatchConditionRequired } from "../requirement";
+import { Section } from "../section";
 import { createDiagnostic } from "../util";
 
 const rule: RelatedSettingsRule = {
